Safari-611.1.10/LayoutTests/fast/canvas/canvas-alphaImageData-behavior-expected.txt

Series of tests to ensure correct behaviour of getImageData and putImageData when alpha is involved

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS imgdata[7] is 254
PASS imgdata[4] is 0
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS imgdata[7] is 254
PASS successfullyParsed is true

TEST COMPLETE
